backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
langchain==0.2.0
langchain-openai==0.1.7
langchain-community==0.2.0
langgraph==0.1.5
openai==1.30.0
pinecone-client==3.2.2
chromadb==0.5.0
cohere==5.5.0
pymupdf==1.24.3
unstructured[pdf]==0.14.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
greenlet==3.0.3
redis==5.0.4
celery==5.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic-settings==2.2.1
httpx==0.27.0
pytest==8.2.0
asyncpg==0.29.0
aiosqlite==0.20.0
email-validator==2.1.1
reportlab==4.2.0
sse-starlette==2.1.0

bcrypt==4.1.3
